variable sample rate. default rate is 48000Hz
the higher sample rate is better for POKEY emulation, particularly for
channels set to 1.79Mhz

+++ gui/gui.go
@@ -32,7 +32,7 @@ type AudioReader interface {
 }
 
 type AudioSetup struct {
-	Freq float64
+	Freq int
 	Read AudioReader
 }
